Vector changed to Float

+++ src/main/java/pl/qus/maxvector/hibernate/customtypes/EVector.kt
@@ -1,6 +1,6 @@
 package pl.qus.maxvector.hibernate.customtypes
 
-class EVector(var points: MutableList<Double> = mutableListOf()) {
+class EVector(var points: MutableList<Float> = mutableListOf()) {
     override fun toString(): String = points.joinToString(",","[","]")
 
     companion object {
@@ -8,7 +8,7 @@ class EVector(var points: MutableList<Double> = mutableListOf()) {
         fun from(value: String): EVector {
             val noBrackets = value.substring(1, value.length - 1)
             val newVector = EVector()
-            newVector.points = noBrackets.split(",").map { it.toDouble() }.toMutableList()
+            newVector.points = noBrackets.split(",").map { it.toFloat() }.toMutableList()
             return newVector
         }
     }
